rename context.tsx to types.ts
