--- frontend/simple-mercari-web/src/common/entity.tsx
@@ -1,13 +1,0 @@
-
-
-export interface Item {
-  id: number;
-  name: string;
-  category_id: number;
-  category_name: string;
-  user_id: number;
-  price: number;
-  description: string;
-  status: number;
-}
-
