+++ frontend/simple-mercari-web/src/common/types.ts
@@ -1,0 +1,2 @@
+
+
